Refactor: Adopt unified repository structure
This commit refactors the application into a more scalable and organized structure as per the new architecture guidelines.

The monolithic `backend/main.py` has been broken down into several components:
- `app/main.py`: Lean FastAPI app instantiation.
- `app/api/routes.py`: API endpoint definitions.
- `app/agent/agent.py`: LangChain agent and executor logic.
- `app/agent/tools.py`: Agent tool definitions.
- `app/core/config.py`: Application configuration.

Additionally, this change:
- Establishes a `tests/` directory with initial tests for the API and tools.
- Updates `requirements.txt` with necessary testing dependencies.
- Ensures the original custom prompt for the agent is preserved to maintain its designed behavior.
- Removes the old `backend/` directory.
