+++ app/core/config.py
@@ -1,0 +1,2 @@
+A1111_URL = "http://127.0.0.1:7860"
+OUTPUT_DIR = "final_outputs"
